Add TransformFunctionTest
