+++ pulsar-functions/transforms/src/main/java/org/apache/pulsar/functions/transforms/TransformFunction.java
@@ -3,6 +3,7 @@ package org.apache.pulsar.functions.transforms;
 import com.google.gson.Gson;
 import com.google.gson.reflect.TypeToken;
 import java.util.ArrayList;
+import java.util.Arrays;
 import java.util.LinkedList;
 import java.util.List;
 import java.util.Map;
@@ -33,7 +34,7 @@ public class TransformFunction implements Function<GenericObject, Void> {
         for (Map<String, String> step : stepsConfig) {
             String type = step.get("type");
             if ("drop-fields".equals(type)) {
-                steps.add(RemoveFieldFunction.of(step));
+                steps.add(newRemoveFieldFunction(step));
             } else {
                 throw new IllegalArgumentException("invalid step type: " + type);
             }
@@ -59,5 +60,23 @@ public class TransformFunction implements Function<GenericObject, Void> {
         }
         transformContext.send();
         return null;
+    }
+
+    public static RemoveFieldFunction newRemoveFieldFunction(Map<String, String> step) {
+        String fields = step.get("fields");
+        if (fields == null || fields.isEmpty()) {
+            throw new IllegalArgumentException("missing required 'fields' parameter");
+        }
+        List<String> fieldList = Arrays.asList(fields.split(","));
+        String part = step.get("part");
+        if (part == null) {
+            return new RemoveFieldFunction(fieldList, fieldList);
+        } else if (part.equals("key")) {
+            return new RemoveFieldFunction(fieldList, new ArrayList<>());
+        } else if (part.equals("value")) {
+            return new RemoveFieldFunction(new ArrayList<>(), fieldList);
+        } else {
+            throw new IllegalArgumentException("invalid 'part' parameter: " + part);
+        }
     }
 }
